test: split workload auth compatibility coverage
