+++ tests/lib/x509-workload-identity.test.ts
@@ -3,8 +3,6 @@ import { expect, vi } from 'vitest';
 import OpenAI, { APIConnectionTimeoutError, APIUserAbortError } from 'openai';
 import { CursorPage } from 'openai/core/pagination';
 import type { RequestInit } from 'openai/internal/builtin-types';
-import type { NullableHeaders } from 'openai/internal/headers';
-import type { FinalRequestOptions } from 'openai/internal/request-options';
 
 const x509Identity = {
   type: 'x509' as const,
@@ -39,19 +37,7 @@ async function* nonReplayableBody() {
 }
 
 class RequestMutatingOpenAI extends OpenAI {
-  authOptionsMutation?: (options: FinalRequestOptions) => void;
-  authHeadersMutation?: (headers: NullableHeaders | undefined) => void;
   requestMutation?: (request: RequestInit) => void;
-
-  protected override async authHeaders(
-    opts: FinalRequestOptions,
-    schemes?: { bearerAuth?: boolean; adminAPIKeyAuth?: boolean },
-  ): Promise<NullableHeaders | undefined> {
-    this.authOptionsMutation?.(opts);
-    const headers = await super.authHeaders(opts, schemes);
-    this.authHeadersMutation?.(headers);
-    return headers;
-  }
 
   protected override async prepareRequest(request: RequestInit): Promise<void> {
     this.requestMutation?.(request);
@@ -674,88 +660,6 @@ describe('OpenAI with X.509 workload identity', () => {
     expect(apiCount).toBe(2);
     expect(bodies).toEqual(['same-body', 'same-body']);
     expect(authorizations).toEqual(['Bearer token-1', 'Bearer token-2']);
-  });
-
-  test('preserves workload-token provenance through a legacy authHeaders override', async () => {
-    let exchangeCount = 0;
-    let apiCount = 0;
-    const apiAuthorizations: (string | null)[] = [];
-    const client = new RequestMutatingOpenAI({
-      apiKey: null,
-      workloadIdentity: x509Identity,
-      fetch: vi.fn(async (url: string | URL | Request, init?: RequestInit) => {
-        if (url.toString().includes('/oauth/token')) {
-          exchangeCount += 1;
-          return tokenResponse(`token-${exchangeCount}`);
-        }
-        apiCount += 1;
-        apiAuthorizations.push(new Headers(init?.headers).get('Authorization'));
-        return apiCount === 1
-          ? Response.json({ error: { message: 'Unauthorized' } }, { status: 401 })
-          : Response.json({ data: [] });
-      }),
-      maxRetries: 0,
-    });
-
-    await expect(client.models.list()).resolves.toMatchObject({ data: [] });
-
-    expect(exchangeCount).toBe(2);
-    expect(apiCount).toBe(2);
-    expect(apiAuthorizations).toEqual(['Bearer token-1', 'Bearer token-2']);
-  });
-
-  test('preserves headers contributed by mutating options in a legacy authHeaders override', async () => {
-    let exchangeCount = 0;
-    let apiCount = 0;
-    const client = new RequestMutatingOpenAI({
-      apiKey: null,
-      workloadIdentity: x509Identity,
-      fetch: vi.fn(async (url: string | URL | Request, init?: RequestInit) => {
-        if (url.toString().includes('/oauth/token')) {
-          exchangeCount += 1;
-          return tokenResponse('workload-token');
-        }
-        apiCount += 1;
-        expect(new Headers(init?.headers).get('Authorization')).toBe('Bearer legacy-custom-token');
-        expect(new Headers(init?.headers).get('X-Legacy-Auth-Hook')).toBe('present');
-        return Response.json({ error: { message: 'Unauthorized' } }, { status: 401 });
-      }),
-    });
-    client.authOptionsMutation = (options) => {
-      options.headers = {
-        Authorization: 'Bearer legacy-custom-token',
-        'X-Legacy-Auth-Hook': 'present',
-      };
-    };
-
-    await expect(client.models.list()).rejects.toMatchObject({ status: 401 });
-    expect(exchangeCount).toBe(1);
-    expect(apiCount).toBe(1);
-  });
-
-  test('drops workload provenance when a legacy authHeaders override replaces the credential', async () => {
-    let exchangeCount = 0;
-    let apiCount = 0;
-    const client = new RequestMutatingOpenAI({
-      apiKey: null,
-      workloadIdentity: x509Identity,
-      fetch: vi.fn(async (url: string | URL | Request, init?: RequestInit) => {
-        if (url.toString().includes('/oauth/token')) {
-          exchangeCount += 1;
-          return tokenResponse('workload-token');
-        }
-        apiCount += 1;
-        expect(new Headers(init?.headers).get('Authorization')).toBe('Bearer legacy-custom-token');
-        return Response.json({ error: { message: 'Unauthorized' } }, { status: 401 });
-      }),
-    });
-    client.authHeadersMutation = (headers) => {
-      headers?.values.set('Authorization', 'Bearer legacy-custom-token');
-    };
-
-    await expect(client.models.list()).rejects.toMatchObject({ status: 401 });
-    expect(exchangeCount).toBe(1);
-    expect(apiCount).toBe(1);
   });
 
   test('retries a replayable 401 when response cleanup fails', async () => {
